change count to long

+++ complitex/complitex-admin/src/main/java/org/complitex/admin/web/UserList.java
@@ -162,7 +162,7 @@ public class UserList extends ScrollListPage {
             }
 
             @Override
-            protected int getSize() {
+            protected Long getSize() {
                 return userBean.getUsersCount(filterModel.getObject());
             }
         };
