+++ complitex/complitex-correction/src/main/java/org/complitex/correction/web/ApartmentCorrectionList.java
@@ -5,14 +5,14 @@ import org.apache.wicket.model.StringResourceModel;
 import org.complitex.address.strategy.building.BuildingStrategy;
 import org.complitex.address.strategy.building.entity.Building;
 import org.complitex.address.util.AddressRenderer;
-import org.complitex.correction.entity.ApartmentCorrection;
-import org.complitex.correction.service.AddressCorrectionBean;
 import org.complitex.common.entity.Correction;
 import org.complitex.common.entity.DomainObject;
 import org.complitex.common.entity.FilterWrapper;
 import org.complitex.common.service.SessionBean;
 import org.complitex.common.strategy.IStrategy;
 import org.complitex.common.web.component.search.SearchComponentState;
+import org.complitex.correction.entity.ApartmentCorrection;
+import org.complitex.correction.service.AddressCorrectionBean;
 
 import javax.ejb.EJB;
 import java.util.List;
@@ -80,7 +80,7 @@ public class ApartmentCorrectionList extends AddressCorrectionList<ApartmentCorr
     }
 
     @Override
-    protected Integer getCorrectionsCount(FilterWrapper<ApartmentCorrection> filterWrapper) {
+    protected Long getCorrectionsCount(FilterWrapper<ApartmentCorrection> filterWrapper) {
         return addressCorrectionBean.getApartmentCorrectionsCount(filterWrapper);
     }
