+++ complitex/complitex-correction/src/main/java/org/complitex/correction/web/StreetTypeCorrectionList.java
@@ -5,12 +5,12 @@ import org.apache.wicket.model.IModel;
 import org.apache.wicket.model.StringResourceModel;
 import org.apache.wicket.request.mapper.parameter.PageParameters;
 import org.complitex.address.strategy.street_type.StreetTypeStrategy;
-import org.complitex.correction.entity.StreetTypeCorrection;
-import org.complitex.correction.service.AddressCorrectionBean;
 import org.complitex.common.entity.Correction;
 import org.complitex.common.entity.DomainObject;
 import org.complitex.common.entity.FilterWrapper;
 import org.complitex.common.service.SessionBean;
+import org.complitex.correction.entity.StreetTypeCorrection;
+import org.complitex.correction.service.AddressCorrectionBean;
 
 import javax.ejb.EJB;
 import java.util.List;
@@ -61,7 +61,7 @@ public class StreetTypeCorrectionList extends AbstractCorrectionList<StreetTypeC
     }
 
     @Override
-    protected Integer getCorrectionsCount(FilterWrapper<StreetTypeCorrection> filterWrapper) {
+    protected Long getCorrectionsCount(FilterWrapper<StreetTypeCorrection> filterWrapper) {
         return addressCorrectionBean.getStreetTypeCorrectionsCount(filterWrapper);
     }
