+++ eirc/eirc-registry/src/main/java/ru/flexpay/eirc/registry/web/component/ContainerListPanel.java
@@ -42,8 +42,8 @@ public class ContainerListPanel<T> extends Panel {
             }
 
             @Override
-            protected int getSize() {
-                return containers.size();
+            protected Long getSize() {
+                return Long.valueOf(containers.size());
             }
         };
