+++ osznconnection/osznconnection-file-handling/src/main/java/org/complitex/osznconnection/file/web/component/account/AccountNumberPickerPanel.java
@@ -149,8 +149,8 @@ public class AccountNumberPickerPanel extends Panel {
             }
 
             @Override
-            protected int getSize() {
-                return getList().size();
+            protected Long getSize() {
+                return Long.valueOf(getList().size());
             }
         };
