+++ osznconnection/osznconnection-file-handling/src/main/java/org/complitex/osznconnection/file/web/pages/account/PersonAccountList.java
@@ -11,7 +11,9 @@ import org.apache.wicket.markup.html.form.Form;
 import org.apache.wicket.markup.html.form.TextField;
 import org.apache.wicket.markup.repeater.Item;
 import org.apache.wicket.markup.repeater.data.DataView;
-import org.apache.wicket.model.*;
+import org.apache.wicket.model.IModel;
+import org.apache.wicket.model.PropertyModel;
+import org.apache.wicket.model.ResourceModel;
 import org.apache.wicket.request.mapper.parameter.PageParameters;
 import org.apache.wicket.util.string.Strings;
 import org.complitex.common.entity.DomainObject;
@@ -89,7 +91,7 @@ public class PersonAccountList extends ScrollListPage {
             }
 
             @Override
-            protected int getSize() {
+            protected Long getSize() {
                 return personAccountBean.getPersonAccountsCount(filterModel.getObject());
             }
         };
