+++ osznconnection/osznconnection-file-handling/src/main/java/org/complitex/osznconnection/file/web/pages/facility/FacilityTarifList.java
@@ -38,7 +38,7 @@ public class FacilityTarifList extends ListTemplatePage<FacilityTarif> {
     }
 
     @Override
-    protected int getCount(FilterWrapper<FacilityTarif> filterWrapper) {
+    protected Long getCount(FilterWrapper<FacilityTarif> filterWrapper) {
         return facilityReferenceBookBean.getFacilityTarifsCount(filterWrapper);
     }
